requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
watchdog==3.0.0
sentence-transformers==2.2.2
chromadb==0.4.22
pyyaml==6.0.1 